Merged emmas into jenny
merged our code, removed the code from main activity and only using the code from gameactivity.
Commented out the last if/else in the popluatenode method because they made the game crash.
Starting the app will lead you to the startscreen with different options and when pressed play on easymode you come to the game